seperating out a bunch of code from functionsDeployHelper
